+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.4.1
+
+
+## Improvements
+
+- Native Android agent updated to version 7.6.0
+- Native iOS agent updated to version 7.5.2
+- Bug fixes for Log Attributes method
+
 ## 1.4.0
 
 ## New Features